requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.0
matplotlib>=3.5.0
tqdm>=4.64.0
pillow>=9.2.0
opencv-python>=4.6.0
scikit-learn>=1.1.0
h5py>=3.7.0
pyyaml>=6.0
openai>=1.1.0
azure-identity>=1.11.0
azure-mgmt-cognitiveservices>=13.0.0 